Update Comment Model and commentController

+++ backend/controllers/commentController.js
@@ -1,16 +1,23 @@
 import asyncHandler from "express-async-handler";
 import Comment from "../models/commentModel.js";
+import Post from "../models/postModel.js";
 import mongoose from "mongoose";
 
 //Route:     POST /new
 //Creates a comment related to a post
 const createComment = asyncHandler(async (req, res) => {
   const user = req.user
-  const { caption } = req.body
+  const { caption, postId } = req.body
+
+  if (!mongoose.Types.ObjectId.isValid(postId)) {
+    res.status(400)
+    throw new Error('Not a valid ID parameter')
+  }
 
   const comment = await Comment.create({
     caption,
-    user
+    user,
+    postId
   });
 
   if (comment) {
@@ -66,19 +73,30 @@ const deleteComment = asyncHandler(async (req, res) => {
   }
 
   const comment = await Comment.findOne({ _id: id });
+  
+  if (!comment) {
+    res.status(400);
+    throw new Error('Error fetching comment');
+  }
 
   if (comment.user._id.toString() !== user._id.toString()) {
     res.status(400)
-    throw new Error('Cannot delete a comment if user did not post it')
+    throw new Error('Cannot delete a comment if user did not post it');
   }
 
-  const deletedComment = await Comment.deleteOne({ _id: id })
+  const postId = comment.postId;
+  const deletedComment = await Comment.deleteOne({ _id: id });
+  const updatedPost = await Post.findOneAndUpdate(
+    { _id: postId },
+    { $pull: { comments: id } },
+    { new: true }
+  );
 
-  if(deletedComment) {
-    res.status(200).json({ message: 'Comment successfully deleted' })
+  if (deletedComment && updatedPost) {
+    res.status(200).json({ message: 'Comment successfully deleted' });
   } else {
     res.status(400)
-    throw new Error('Error - Unable to delete this comment')
+    throw new Error('Error - Unable to delete this comment');
   }
 });
